aula2/requirements.txt
dash
dash-mantine-components==0.11.0
dash-bootstrap-components
gunicorn
pandas
dash-chart-editor
openai
dash-ag-grid
redis
sqlalchemy
langchain_community
langchain_openai
langchain-experimental==0.3.4
langchain-google-genai 
google-generativeai
python-dotenv
psycopg2
loguru==0.7.3